\name{GeoRiskR-package}
\Rdversion{2.15}
\alias{GeoRiskR-package}
\alias{GeoRiskR}
\docType{package}
\title{GeoRisk-package}
\description{Shear strength parameters of soils, rocks, geosynthetics, some useful functions for multivariate analysis and reliability analysis in geotechnical engineering}
\details{
	\tabular{ll}{
		Package: \tab GeoRiskR\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2013-5-15\cr
		License: \tab GNU 2.15 or later\cr
		LazyLoad: \tab yes\cr
	}
}
\author{
	Xing Zheng Wu
	Maintainer:  <xingzhengwu@gmail.com>
}
\references{
	Wu XZ. 2013. Implementing reproducible data analysis and statistical computing for the geosciences in R. Computers and Geosciences. in press
}
\keyword{ package }
	\seealso{
}
\examples{
	## listing the data for soils
	SoilShear
	## listing the data for rocks
	RockShear
	## listing the data for geosynthetic
	GCLShear
	## listing the data for soils by Mousavi et al. 2011
	MousaviSet
}
